feat(wasm): split os.Chmod and os.SameFile into _native/_wasm stubs
tinygo's wasm runtime omits os.Chmod and os.SameFile. Moves chmodFn
and sameFile out of githooks.go into build-tagged pairs so the WASM
build compiles cleanly without those symbols. Same pattern applied to
fix.atomicWriteFile via fix/chmod_native.go + fix/chmod_wasm.go.

Also splits schema.WriteIndex / schema.ValidateIndex (OS filesystem
reads and writes) into index_native.go (native) and index_wasm.go
(no-op stubs), leaving the pure in-memory BuildIndex in index.go for
both build targets.
